Create an oxidized mixer
Create an oxidized cubeb-mixer, which is adapted from cubeb_mixer.cpp,
as the new mixer module. The new mixer is able to mix from any input
channels to any output channels, without predefined channel order.

The old cubeb-mixer has a strict rule for channel order. For example,
for an stereo layout, it must be 2-channel. One is front-left, the other
is front-right. And the order of the channels must be [front-left,
front-right]. However, in MacOS, the N-channel output device can actually
be set to an 1-to-N output device with any channel order. For example,
the 6-channel output device can be set as a stereo with channel order:
[unused, unused, front-left, unsued, front-right, unused], where the
unused channel should be silience.

With the new mixer, the restriction of the channel order would be loosen
and audio is able to be mixed from any channel layout to any channel
layout, without any pre-defined channel order and limitation of
channels' number.